services/user/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
greenlet==3.0.3

# Database & ORM
alembic==1.13.1
SQLAlchemy==2.0.29
asyncpg>=0.29,<1.0
psycopg2-binary==2.9.9

# Auth & Security
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# .env loader
python-dotenv==1.0.1

# Settings & Validation
pydantic[email]==2.6.4
pydantic-settings>=2.0,<3.0
